feat: use Session Provider client side and check server side with await auth
