+++ src/app/dashboard/layout.tsx
@@ -1,5 +1,18 @@
-import { UserProvider } from "@/providers/UserProvider";
+import { redirect } from "next/navigation";
 
-export default function DashboardLayout({ children }: { children: React.ReactNode }) {
-	return <UserProvider>{children}</UserProvider>;
+import { AppLayout } from "@/components/layouts/AppLayout";
+import { auth } from "@/lib/auth";
+
+export default async function DashboardLayout({
+	children,
+}: {
+	children: React.ReactNode;
+}) {
+	const session = await auth();
+
+	if (!session) {
+		redirect("/auth/login");
+	}
+
+	return <AppLayout>{children}</AppLayout>;
 }
